+++ client/src/components/Header.tsx
@@ -1,7 +1,5 @@
-import logo_atrium from '../assets/svg/atrium-logo.svg';
-import logo_login from '../assets/svg/user-icon.svg';
-
-type TNavItem = { label: string; href: string };
+import { SectionLogo, SectionLogin, SectionNavItem } from './elements/containers';
+import { TNavItem } from './utils/types';
 
 export default function Header() {
   const list_nav_items: TNavItem[] = [
@@ -13,42 +11,9 @@ export default function Header() {
   return (
     <header className="sticky top-0 z-100">
       <nav className="bg-teal-800 border-b border-gray-200 shadow-sm w-full pt-2 relative">
-        {/* Absolute -  Login Button */}
-        <a
-          href="/login"
-          className="absolute right-8 top-8 transform px-4 h-8 rounded-3xl bg-slate-300 flex items-center justify-left cursor-pointer hover:bg-slate-100 hover:text-teal-800 transition duration-500 active:scale-95"
-        >
-          <img src={logo_login} alt="User" width="24" height="24" className="text-teal-800" />
-          <span className="px-2 text-sm text-teal-800">Login</span>
-        </a>
-        {/*  Relative - Main Navigation */}
-        <div className="flex items-center justify-center">
-          {/* Static - Large Logo Banner */}
-          <a className="flex active:scale-98 justify-center items-center p-4 cursor-pointer">
-            <img src={logo_atrium} alt="Atrium" width="48" height="48" />
-            <h1 className="text-2xl font-bold text-slate-200 font-serif px-2 hover:text-white transition duration-900">
-              Atrium
-            </h1>
-          </a>
-        </div>
-        {/* Static - Nav Items Container */}
-        <div className="flex items-center justify-center">
-          <ul className="flex justify-center h-8">
-            {list_nav_items.map((item: TNavItem) => (
-              <li
-                key={item.label}
-                className="flex items-center px-8 py-2 text-l text-slate-300 font-serif font-bold cursor-pointer hover:border-b-6 hover:border-teal-600 transition duration-900"
-              >
-                <a
-                  href={item.href}
-                  className="inline-block hover:brightness-200 hover:text-white transition duration-900 active:scale-90"
-                >
-                  {item.label}
-                </a>
-              </li>
-            ))}
-          </ul>
-        </div>
+        <SectionLogin />
+        <SectionLogo />
+        <SectionNavItem nav_items={list_nav_items} />
       </nav>
     </header>
   );